refactor: move worker to worker/site-entry.ts; wrangler.toml -> wrangler.jsonc
