--- wrangler.toml
@@ -1,24 +1,0 @@
-name = "pulletsforever-com"
-compatibility_date = "2026-02-18"
-main = "workers/site/index.ts"
-upload_source_maps = true
-
-[assets]
-directory = "./dist"
-binding = "ASSETS"
-run_worker_first = true
-
-[observability]
-enabled = false
-head_sampling_rate = 1
-
-[observability.logs]
-enabled = true
-head_sampling_rate = 1
-persist = true
-invocation_logs = true
-
-[observability.traces]
-enabled = true
-persist = true
-head_sampling_rate = 1
